Fix for points-less player messing up Dealer cache
The Dealer is responsible for knowing about 2 good players before
passing through their plays - but can't be responsible for how the game
decides to vet a player, the game needs to decide that, if it even wants
to. (Stockpile doesn't, Reverse does, for example).

Before this change, things would get out of whack.