Dynamic command system pushed to games.
The Dealer no longer has hardcoded methods that are the sum total of
commands across all games, but rather passes along any game command to
the game, the BaseGame has a sendCommand which depends on the subclass
to implement an array of command info (regex and matching methods).

Hubot dealer just registers a catch-all regex that goes down to the
dealer, the dealer down to the game, if the game doesn't respond then
it's ignored.

Two things are broken as of this - the play cache for Reverse Hold 'Em,
and funding players ... and AI is broken, too. So - three.