Fixed problem with cache and next round guard.
After putting back the play cache, attempting a play during the 'wait
for next round' phase would just tuck the play into the cache saying
'hey, we need another player first.' If a second player played, then the
cache was dumped into the game proper, which would then blow up because
the next round hadn't started yet.

:|

This also enhances the AI commands in Reverse Hold 'Em to push the
available commands into the various play state classes, so the AI is
'smarter' in that it will no longer attempt a play when it's not
allowed.