CardPoker hand comparisons working properly.
All of the non-grouping hands merely sort by rank to compare hands of
the same type.

Grouping hands sort the hands by weighting the groups of ranks first -
if there are 3 of a rank, then those are a higher priority than 2 of the
same rank, then within groups, cards are sorted by high rank. Cards not
participating in groups are left in to account for kicker calculation.
Kickers may be optional across different game types, but for now they
are hardcoded in since Texas Hold Em uses them.