First shot at AiBetBrain.
This is very simple implementation so far, that only passes 3 of 4
tests. With only 2 cards dealt, finding all matches with the current
HandMatcher will identify a pair, a starter flush or garbage, but it
can't identify a possible straight.

This analysis is also going to require the number of remaining cards,
because if I have 3 cards of the same suit, but only one more community
card to turn, then it's no longer a possible flush.

So ... much to think about :)